requirements.txt
google-generativeai>=0.8.0
openai>=1.0.0
anthropic>=0.34.0

